初步完成simple_loader加载器功能

+++ src/include/comm.h
@@ -32,8 +32,8 @@ SOFTWARE.
 #define ALIGN2PGSZ(addr,T) (addr & (~(T)(ENTRY_SIZE -1)))
 #define MOD2PGSZ(addr,T) (addr & (T)(ENTRY_SIZE -1))
 
-#define ROUND(x, align)  (((uintptr_t)x) & ~(align - 1))
-#define MOD(x, align)    (((uintptr_t)x) & (align - 1))
+#define ROUND(x, align)  (((uint64_t)x) & ~((uint64_t)align - 1))
+#define MOD(x, align)    (((uint64_t)x) & ((uint64_t)align - 1))
 #define push(sp, T, ...) ({ *((T*)sp) = (T)__VA_ARGS__; \
                             sp = (void *)((uintptr_t)(sp) + sizeof(T)); })
